Revert "Style:sort버튼 위치 변경"
This reverts commit eed8870b269eddd7931661f3e5c200d902de64be.
