+++ src/components/MainPage/RecentPostListWrap.tsx
@@ -6,9 +6,5 @@ export const RecentPostListwrap = () => {
     queryKey: ["posts"],
     queryFn: () => getPosts({ page: 1, limit: 3 })
   });
-  return (
-    <div className="py-2 md:px-4">
-      <PostList posts={data.posts} />
-    </div>
-  );
+  return <PostList posts={data.posts} />;
 };
